Fixed the retrieving issue for the weather stations

+++ WeatherImageGenerator.Domain/Entities/Weather/WeatherStation.cs
@@ -10,6 +10,6 @@
         string WeatherDescription,
         decimal Temperature,
         decimal FeelTemperature,
-        int WindSpeed,
+        decimal WindSpeed,
         string WindDirection);
 }
